+++ contactos.py
@@ -1,14 +1,11 @@
 #=======IMPORTACIONES=============================================
 import os
-# import tkinter as tk
 from tkinter import *
 from tkinter import messagebox
 from PIL import ImageTk, Image
 from PIL import *
 import ast
-#import contactos 
-
-
+import ventanaPrincipal as VP
 
 root = Tk()
 root.title('Login')
@@ -20,14 +17,11 @@ dir_carpetas = os.path.dirname(__file__)
 img_carpeta = os.path.join(dir_carpetas,"imagenes")
 root.iconbitmap(os.path.join(img_carpeta,"contactos.ico"))
 
-img_carpeta = os.path.join(dir_carpetas,"imagenes")
-root.iconbitmap(os.path.join(img_carpeta,"contactos.ico"))
-
 def singin():#Esta es la funcion de inicio de sesion
     username=user.get()#Tomamos los valores ingresados del usuario
     password=code.get()#Tomamos los valores ingresados de la contraseña
 
-    file=open('INFO2024_Contactos/Datos_login_Usuario/user_data.txt','r')
+    file=open('user_data.txt','r')
     d=file.read()
     r=ast.literal_eval(d)
     file.close()
@@ -37,75 +31,17 @@ def singin():#Esta es la funcion de inicio de sesion
 
 #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ Parte Del Codigo de Contactos @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 
-
     if username in r.keys() and password==r[username]:
         root.destroy()
-        
-
-        #======FUNCIONES==============================================
-        def funcion_login():
-            messagebox.showwarning("Login", "Aquí se hace Inicio de Sesión")
-
-        def funcion_logout():
-            ventana.destroy()
-            # messagebox.showwarning("Logout", "Aqui se hace CERRAR sesión")
-
-        def mostrar_contactos():
-            messagebox.showwarning("Contactos", "Aquí se muestra todos los contactos, y los botones Crear/Eliminar/Editar") 
-
-        def funcion_exportar():
-            messagebox.showwarning("Exportar", "Aquí la opción de Exportar los contactos a EXCEL")
-
-
-        #=====PROGRAMA PRINCIPAL===============================================
-        ventana = Tk()
-        lista_contactos = Listbox(ventana)
-        login = False
-
-        dir_carpetas = os.path.dirname(__file__)
-        img_carpeta = os.path.join(dir_carpetas,"imagenes")
-        ventana.iconbitmap(os.path.join(img_carpeta,"contactos.ico"))
-
-        ventana.title('Agenda Personal')
-        ventana.geometry('800x600')
-        ventana.configure(bg="#4B6587")
-
-        img_logo_ppal = ImageTk.PhotoImage(Image.open(os.path.join(img_carpeta,"contactos.png")).resize((100,100)))
-        img_mostrar = Label(image=img_logo_ppal)
-        img_mostrar.place(relx=0.8,rely=0.8)
-
-        barra_menu = Menu(ventana)
-        ventana.config(menu=barra_menu)
-        menu_principal = Menu(barra_menu)
-        barra_menu.add_cascade(label ='Menú', menu=menu_principal)
-        submenu = Menu(menu_principal)
-
-        if(login):
-            menu_principal.add_cascade(label = 'Login', command=funcion_login)
-        else:
-            menu_principal.add_cascade(label = 'Contactos', menu=submenu)
-            menu_principal.add_cascade(label = 'Logout', command=funcion_logout)
-
-        submenu.add_cascade(label = 'Actualizar Contactos',command=mostrar_contactos)
-        submenu.add_cascade(label = 'Exportar a Excel', command=funcion_exportar) 
-
-        ventana.mainloop()
-
+        VP.ventanaPrincipal()
+    
     #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ Fin de La Parte del Codigo de contactos @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
-
 
     else:
         messagebox.showerror('Inválido','Usuario o contraseña incorrecta')
 
-'''
-Deje como principal ejecucion el login, para que el usuario antes de entrar a guaradar sus contactos tenga
-que hacer un inicio de sesion previo o bien en el caso de que no tenga una cuenta, tiene que registarse
-para poder ingresar a crear contactos
-'''
-
 
 ######/////////////////////En esta parte va a estar el codigo donde vamos a ejecutar la ventana del registro de sesion
-
 
 def singup_command():
 
@@ -132,20 +68,18 @@ def singup_command():
                 file.truncate(0)
                 file.close()
 
-                file=open('INFO2024_Contactos/user_data.txt', 'w')
+                file=open('user_data.txt', 'w')
                 w=file.write(str(r))
 
                 messagebox.showinfo('Registro','Registro de cuenta exitosa')
                 windows.destroy()
 
             except:
-                file=open('INFO2024_Contactos/user_data.txt', 'w')
-                pp=str({'Usuario':'contraseña'})
+                file=open('user_data.txt', 'w')
+                pp=str({'Usuario':'contrasenia'})
                 file.write(pp)
                 file.close()
                 
-
-
       else:
          messagebox.showerror('Invalido',"Ambas contraseñas deben coincidir")
 
@@ -153,11 +87,9 @@ def singup_command():
         windows.destroy()
 
 
+    img = PhotoImage(Image.open(os.path.join(img_carpeta,"logup.png")))
 
-
-
-    img = PhotoImage(file='INFO2024_Contactos/img_login_out/logup.png')
-    Label(windows,image=img,border=0,bg='white',).place(x=50,y=90)
+    Label(windows,image=img,border=0,bg='white').place(x=50,y=90)
 
     frame = Frame(windows,width=350,height=390,bg='#fff')
     frame.place(x=480,y=50)
@@ -166,7 +98,6 @@ def singup_command():
     heading.place(x=100,y=5)
 
     ####------------------------------------------------------------
-
     def on_enter(e):
         user.delete(0,'end')
     def on_leave(e):
@@ -182,8 +113,6 @@ def singup_command():
 
     Frame(frame,width=295,height=2,bg='black').place(x=25,y=107)
 
-
-
     ####--------------------------------------------------
 
     def on_enter(e):
@@ -191,7 +120,6 @@ def singup_command():
     def on_leave(e):
         if code.get()=='':
             code.insert(0,'Contraseña')
-
 
     code = Entry(frame,width=25,fg='black',border=0,bg='white',font=('Microsoft Yahei UI ligth',11))
     code.place(x=30,y=150)
@@ -202,7 +130,6 @@ def singup_command():
     Frame(frame,width=295,height=2,bg='black').place(x=25,y=177)
 
     ####--------------------------------------------------
-
     def on_enter(e):
         conform_code.delete(0,'end')
     def on_leave(e):
@@ -218,7 +145,6 @@ def singup_command():
     Frame(frame,width=295,height=2,bg='black').place(x=25,y=247)
 
     #---------------------------------------------
-
     Button(frame,width=39,pady=7,text='Inicio de sesion',bg='#57a1f8',fg='white',border=0,command=signup).place(x=35,y=280)
     label = Label(frame,text='Ya tengo una cuenta',fg='black',bg='white',font=('Microsoft Yahei UI ligth',9))
     label.place(x=90,y=340)
@@ -230,17 +156,11 @@ def singup_command():
 
 
 ##################################//////////////////////////////
-
-
-img = PhotoImage(file='INFO2024_Contactos/img_login_out/login.png')
-Label(root, image = img, bg= "white").place(x=50,y=50)
-
 frame=Frame(root, width=350, height=350, bg="white")
 frame.place(x=480,y=70)
 
 heading=Label(frame, text='Sing in', fg='#57a1f8',bg='white',font=('Microsoft YaHei UI Ligth',23,'bold'))
 heading.place(x=100,y=5)
-
 
 
 #------------------------------------------------------------------------------------------------
@@ -261,8 +181,6 @@ user.bind('<FocusOut>', on_leave)
 Frame(frame,width=295,height=2,bg='black').place(x=25,y=107)
 
 
-
-
 ###########################-----------------------------------------------------------------------
 def on_enter(e):
     code.delete(0, 'end')
@@ -281,14 +199,11 @@ code.bind('<FocusOut>', on_leave)
 Frame(frame,width=295,height=2,bg='black').place(x=25,y=177)
 
 #################################################
-
 Button(frame,width=39,pady=7,text='Sing in',bg='#57a1f8',fg='white',border=0, command=singin).place(x=35,y=204)
 label=Label(frame,text="¿No tienes una cuenta?",fg='black',bg='white',font=('Microsoft YaHei UI Ligth',9))
 label.place(x=75,y=270)
 
 sin_cuenta= Button(frame,width=8,text='Inicia sesión',border=0,bg='white',cursor='hand2',fg='#57a1f8',command=singup_command)
 sin_cuenta.place(x=215,y=270)
-
-
 
 root.mainloop()
