+++ user_data.txt
@@ -1,1 +1,1 @@
-{'Usuario': 'contraseña'}
+{'Usuario':'contrasenia','info':'123456'}
